requirements.txt
typer>=0.16.0
PyYAML>=6.0
requests>=2.28.0
